Add 'user' domain type for regular and custom webhook
