+++ pkg/artifactory/resource/webhook/resource_artifactory_webhook_user.go
@@ -1,0 +1,25 @@
+package webhook
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+type EmptyWebhookCriteria struct{}
+
+var userWebhookSchema = func(webhookType string, version int, isCustom bool) map[string]*schema.Schema {
+	return getBaseSchemaByVersion(webhookType, version, isCustom)
+}
+
+var packEmptyCriteria = func(artifactoryCriteria map[string]interface{}) map[string]interface{} {
+	return map[string]interface{}{}
+}
+
+var unpackEmptyCriteria = func(terraformCriteria map[string]interface{}, baseCriteria BaseWebhookCriteria) interface{} {
+	return EmptyWebhookCriteria{}
+}
+
+var emptyCriteriaValidation = func(ctx context.Context, criteria map[string]interface{}) error {
+	return nil
+}
